% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin-importFrom.R
\name{generate_importFrom}
\alias{generate_importFrom}
\title{Generate \code{importFrom} Roxygen Tags}
\usage{
generate_importFrom(script)
}
\arguments{
\item{script}{R script to generate tags for.}
}
\description{
Generate \code{importFrom} Roxygen Tags
}
\seealso{
\href{https://github.com/dreamRs/prefixer}{dreamRs/prefixer}

Other importFrom: 
\code{\link{check_import_from}()},
\code{\link{import_from}()},
\code{\link{tidy_importFrom}()}
}
\concept{importFrom}
