% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin-importFrom.R
\name{import_from}
\alias{import_from}
\title{Generate \code{importFrom} Roxygen Tag from a Function
Works only if functions used are prefixed in the body.}
\usage{
import_from(fun, quiet = FALSE)
}
\arguments{
\item{fun}{A function.}

\item{quiet}{Logical, display output to console ?}
}
\value{
Invisible character string
}
\description{
Generate \code{importFrom} Roxygen Tag from a Function
Works only if functions used are prefixed in the body.
}
\examples{

my_fun <- function(path) {
  utils::read.table(file = path, header = FALSE, sep = "\t")
}
import_from(my_fun)
}
\seealso{
Other importFrom: 
\code{\link{check_import_from}()},
\code{\link{generate_importFrom}()},
\code{\link{tidy_importFrom}()}
}
\concept{importFrom}
