% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jimsaddins-package.R
\docType{package}
\name{jimsaddins-package}
\alias{jimsaddins}
\alias{jimsaddins-package}
\title{jimsaddins: RStudio Addins}
\description{
A collection of RStudio Addins.
}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
