% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin-importFrom.R
\name{process_imports}
\alias{process_imports}
\title{Process Imports}
\usage{
process_imports(imports)
}
\arguments{
\item{imports}{list of \code{importFrom} statements.}
}
\value{
vector of processed import statements
}
\description{
This helper function processes \code{importFrom} roxygen2 tags for use in the
\code{\link[=tidy_importFrom]{tidy_importFrom()}} funciton/add-in.
}
\keyword{internal}
