% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin-importFrom.R
\name{tidy_importFrom}
\alias{tidy_importFrom}
\title{Tidy importFrom Roxygen Statements}
\usage{
tidy_importFrom()
}
\description{
statements, in the \code{R} directory of the current working directory will be
moved to a file \code{R/imports.R}, cleaned and sorted.
}
\examples{
\dontrun{
tidy_importFrom()
}

}
\seealso{
Other importFrom: 
\code{\link{check_import_from}()},
\code{\link{generate_importFrom}()},
\code{\link{import_from}()}
}
\concept{importFrom}
